12.4 corrected and working, no tests

+++ src/main/java/com/bobi89/medicalclinic/model/entity/appointment/AppointmentRequest.java
@@ -9,5 +9,5 @@ import java.time.LocalDateTime;
 @AllArgsConstructor
 public class AppointmentRequest {
     private LocalDateTime localDateTime;
-    private int durationMinutes;
+    private long durationMinutes;
 }
